add logger

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { v1 as uuidv1 } from "uuid";
 import { CreateTaskDTO } from './dto/create-task.dto';
 import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
@@ -16,16 +17,16 @@ export class TasksService {
 
     }
 
-    async getTasks(filterDto:GetTasksFilterDTO):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto:GetTasksFilterDTO, user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async createTask(createTaskDto:CreateTaskDTO):Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto:CreateTaskDTO, user:User):Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async getTaskById(taskId:number):Promise<Task>{
-        const found = await this.taskRepository.findOne(taskId);
+    async getTaskById(taskId:number, user:User):Promise<Task>{
+        const found = await this.taskRepository.findOne(taskId, {where : {id:taskId, owner:user}});
 
         if(!found){
             throw new NotFoundException(`Task with id ${taskId} not found`)
@@ -34,9 +35,9 @@ export class TasksService {
         return found;
     }
 
-    async deleteTask(taskId : number):Promise<number>{
+    async deleteTask(taskId : number, user:User):Promise<number>{
 
-        const result = await this.taskRepository.delete(taskId)
+        const result = await this.taskRepository.delete({id:taskId, owner:user})
 
         if(result.affected === 0){
             throw new NotFoundException(`Task with id ${taskId} not found`)
@@ -45,8 +46,8 @@ export class TasksService {
         return taskId;
     }
 
-    async updateTask(taskId:number, status:TaskStatus):Promise<Task>{
-        const task = await this.getTaskById(taskId);
+    async updateTask(taskId:number, status:TaskStatus, user:User):Promise<Task>{
+        const task = await this.getTaskById(taskId, user);
 
         task.status = status;
         await task.save();
